fix: シャッターボタンのレスポンスを改善
