+++ src/layouts/Today/ResultView/index.tsx
@@ -5,7 +5,7 @@ import stand from './stand.svg'
 import Image from 'next/image'
 import arrow from './arrow.svg'
 import Link from 'next/link'
-import { use, useRef } from 'react'
+import { useRef } from 'react'
 
 export type ResultViewProps = {
   captureData: CaptureData
@@ -15,6 +15,7 @@ const ResultView: React.FC<ResultViewProps> = ({ captureData }) => {
   const { flavor, colors, pattern } = captureData
   const initialTime = useRef<number>(Date.now())
 
+  if (!colors) return <></>
   return (
     <div className={styles['container']}>
       <div className={styles['name-container']}>
